+++ py/better_previews/__init__.py
@@ -1,0 +1,3 @@
+from . import previewer, tae_vid
+
+__all__ = ("previewer", "tae_vid")